fixed review issues rename printf_char_context to printf_context remane printf_wchar_context to wprintf_context
renamed the old printf_context template to basic_printf_context_t. the original wish was to rename it basic_printf_context, but that clashed with the name of the inner typedef. this style matches the format_context_t struct.

try to fix compile errors for older gcc versions by storing a temporary.