Avoid usage of begin/end for member variables in chrono.h
begin and  end are commonly used  (within and outside the  standard) for
member functions. Due to the  naming convensions of this project, common
internal  names such  as  `begin_` and  `m_begin`  are unfavourable  for
structs. This the alternative names  `first` and `last` which are common
for the interface of standard algorithms, have been chosen.

`begin()`  and  `end()`  member  functions  have  been  added  to  those
structs, and the calling code modified; thus increasing consistency with
surrounding code.