Avoid usage of begin/end for member variables in format.h
As with parent commit, begin and  end are commonly member functions, not
member variables. To  fit in with the style of  the project, `first` and
`last` have been used as an alternative within the `arg_join` struct.

`begin()` and  `end()` member functions  have been added to  the struct,
and  the  calling  code   modified;  thus  increasing  consistency  with
surrounding code.