Allow for different iterator types for arg_join
With  C++20,  and already  in  the  ranges-v3  library (and  many  other
projects); the iterators that can  be compared within algorithms may not
be the same type.  Even `begin` and `end` iterators may  not be the same
time, but still correspond to the same range.

This patch allows for a "sentinel" type.