Better fix for all compilers
The Nvidia compiler complained about unreachable code with original throw followed by return 0.  However, if the return 0 was removed, then the MSC compiler complained about function not returning value. This was due to the use of `FMT_THROW` which "tricks" the MSC compiler into thinking it is possible for the `FMT_THROW` to return.  

If we just use the `thow format_error(...)` here with no return, then all compilers should be able to deduce that the code will not return following the explicit throw and it should eliminate all warnings from Nvidia, MSC, and other compilers.