Fix clang -Wsign-conversion warning in bigint::subtract_bigits.
subtract_bigits's index parameter is only passed to basic_memory_buffer::operator[] (size_t).

subtract_bigits is only called from subtract_aligned, which passes it 'i' for the index.

subtract_aligned initialises 'i' to 'other.exp_ - exp_', after asserting that other.exp_ >= exp_. As such the result of this subtraction must always be positive, and so can safely be cast to a size_t.

As 'i' is only ever incremented after its assignment it too can be a size_t.