Optimize bloat generated by write_padding
Reimplment write_padded function as a structure that removes align
and <typename F> template parameters in order to reduce the number of
template specializations generated. This results in a binary size
reduction.

Caveat, static_assert to detect invalid alignment has been removed as
the aligment parameter has been removed from the template arguments and
added to the constructor parameters.

Resolves #1774