cmake: treat public includes as system includes to avoid warning from headers
Include directories for target fmt was splitted into INTERFACE and PRIVATE.
INTERFACE includes was marked as SYSTEM to suppress warnings in targets that
are linking fmt:fmt. PRIVATE includes are left as normal includes so that
warnings still occur when building fmt.

Includes for fmt-header-only are now marked as SYSTEM.