Allow chrono to work with defined thousands separator
This allows the 8.1.0 and later chrono.h implementation to work (i.e. do not emit any undefined symbol warnings) if the user has `FMT_STATIC_THOUSANDS_SEPARATOR` defined.  The chrono implementation is relying on this definition of `locale_ref` to be defined.  

Moving the `#ifndef` line to still define `locale_ref` and only select which version of `thousands_sep_impl` gets defined.